Move hook to before create
